% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/b_rcs_prime.R
\name{b_rcs_prime}
\alias{b_rcs_prime}
\title{Derive first derivatives of RCS}
\usage{
b_rcs_prime(x, knots)
}
\arguments{
\item{x}{vector of values}

\item{knots}{vector of knot locations}
}
\description{
function that derives the first derivative of the restricted cubic splines 
for a value/vector of values, given the knots
}
