% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cs.per.R
\name{cs.per}
\alias{cs.per}
\title{Periodic cubic splines design matrix}
\usage{
cs.per(x, knots = NULL, nk = 5, xmax = max(x, na.rm = TRUE),
  xmin = min(x, na.rm = TRUE))
}
\arguments{
\item{x}{numeric vector}

\item{knots}{vector with knot locations}

\item{nk}{number of knots, overridden if knots are provided}

\item{xmax}{value of the end of the cycle}

\item{mmin}{value of the beginning of the cycle}
}
\description{
implementation of the formulas of zhang for any support of x
}
