% Generated by roxygen2 (4.0.1): do not edit by hand
\name{cyclicCubicSplineTPB}
\alias{cyclicCubicSplineTPB}
\title{Function to plot a cyclic cubic spline using truncated power basis}
\usage{
cyclicCubicSplineTPB(data, coef = FALSE)
}
\description{
A cyclic function is plotted using cubic splines with truncated power basis..
# function for calculating the coeffients ####
}

