% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CyclicFractionalPolynomials.R
\name{cyclicFractionalPolynomial}
\alias{cyclicFractionalPolynomial}
\title{Function to calculate coefficients for a cyclic fractional polynomial.}
\usage{
cyclicFractionalPolynomial(data)
}
\arguments{
\item{data}{The data to estimate on. Assumes a column named CyclicTime.}
}
\description{
Returns the  coefficients for a cyclic function constructed using cyclic 
fractional polynomial.
}

