% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/CyclicFractionalPolynomialsPlot.R
\name{cyclicFractionalPolynomialCurve}
\alias{cyclicFractionalPolynomialCurve}
\title{Function for generating the curve estimated by cyclic fractional polynomial.}
\usage{
cyclicFractionalPolynomialCurve(data, curve = FALSE, add = FALSE)
}
\arguments{
\item{data}{Data to estimate on.}

\item{curve}{If TRUE, a curve is plotted instead of points.}

\item{add}{If TRUE, the plot is added to an existing plot.}
}
\description{
Function for generating the curve estimated by cyclic fractional polynomial.
}

