% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/look_at_simulation.R
\name{look_at_simulation}
\alias{look_at_simulation}
\title{Prints ot results for a particular simulation}
\usage{
look_at_simulation(par4sin, par5sin, results_list, file = NULL)
}
\description{
Prints out results for a particular simulation
 (if they exist) saved ina a list.
}

