% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.rcs.mod.R
\name{Plot.rcs.mod}
\alias{Plot.rcs.mod}
\title{Just some ploting function}
\usage{
Plot.rcs.mod(My.mod, my.var, my.ylab = "Response", my.xlab = "Covariate",
  my.xlim = NULL, my.ylim = NULL, my.knots = NULL, my.title = NULL,
  vlines = NULL, hlines = NULL, my.cex.lab = NULL, my.axes = TRUE,
  my.cex.main = NULL, my.cex.axis = NULL)
}
\description{
Plots graph in a certain, nice looking way ;)
}
