% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcs.per.1c.R
\name{rcs.per.1c}
\alias{rcs.per.1c}
\title{RCS.PER, one restriction}
\usage{
rcs.per.1c(x, knots, xmin, xmax)
}
\description{
Experimental: 
periodic restricted cubic spline, with only the constraint on equality of estimated values at beginning/end of period
}
