% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcs_per.R
\name{rcs_per}
\alias{rcs_per}
\title{Generate design matrix for periodic restricted cubic spline}
\usage{
rcs_per(x, knots = NULL, nk = 5, xmin = min(x, na.rm = TRUE),
  xmax = max(x, na.rm = TRUE))
}
\arguments{
\item{x}{numerical x values to transform to new basis}

\item{knots}{vector with locations of the knots of the spline}

\item{nk}{number of knots, used only if the knots are not specified, overridden otherwise}

\item{xmin}{value of the (theoretical) minimum of x}

\item{xmax}{value of the (theoretical) maximum of x}
}
\description{
Generate design matrix for periodic restricted cubic spline.
}
