% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sine_function.R
\name{sine_function}
\alias{sine_function}
\title{Sine function used in the simulations (for simulating probability)}
\usage{
sine_function(par1sin, par2sin, par3sin, par4sin, par5sin)
}
\description{
Sine function used in the simulations (for simulating probability)
}

