% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/CyclicCubicSplines.R
\name{cyclicCubicSplineTPB}
\alias{cyclicCubicSplineTPB}
\title{Function to calculate coefficients for a cyclic cubic spline using truncated power basis.}
\usage{
cyclicCubicSplineTPB(data, coef = FALSE)
}
\description{
The coefficients for a cyclic function constructed using cubic splines with truncated power basis.
}

