% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/CyclicCubicSplines.R
\name{cyclicCubicSplineTPB}
\alias{cyclicCubicSplineTPB}
\title{Function to calculate coefficients for a cyclic cubic spline using
truncated power basis.}
\usage{
cyclicCubicSplineTPB(data, coef = FALSE)
}
\arguments{
\item{data}{The data to work on. It must include a column name CycliTime
that has domain in [0,1].}

\item{coef}{If TRUE returns the coeficients.
Otherwise returns the solution matrix.}
}
\description{
Returns the coefficients for a cyclic function constructed using cubic splines
with truncated power basis.
}

