% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/f.sim.per.splines.R
\name{f.sim.per.splines}
\alias{f.sim.per.splines}
\title{Simulation function}
\usage{
f.sim.per.splines(B = 100, n = 100, n.new = 100, nk = 5, knots = NULL,
  quantiles.cs = c(0.05, 0.25, 0.5, 0.75, 0.95), par1sin = 1,
  par2sin = 0.25, par3sin = 0.25, par4sin = 1, par5sin = 0)
}
\arguments{
\item{B}{number of simulations}

\item{n}{number of points in training sample?}

\item{n.new}{number of points in test sample?}

\item{nk}{number of knots}

\item{knots}{vector with locations of knots}

\item{par1sin}{first parameter of sinusoidal function}

\item{par2sin}{second parameter of sin function}

\item{par3sin}{third parameter of sin function}

\item{par4sin}{the number of sine cycles per one of our periods}

\item{par5sin}{offset of the sine curve in multiples of 2*Pi (defaults to 0); with 0.25 we would get cosine curve?}

\item{quatiles.cs}{quantiles at which to place the knots by default?}
}
\description{
Simulation function
}

