% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sine_function.R
\name{sine_function}
\alias{sine_function}
\title{Sine function used in the simulations (for simulating probability)}
\usage{
sine_function(x, par1sin = 1, par2sin = 0.25, par3sin = 0.25,
  par4sin = 1, par5sin = 0)
}
\description{
Sine function used in the simulations (for simulating probability)
}

